+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- `apple-layout` build feature.
+- `feat` table.
 
 ## [0.14.0] - 2021-12-28
 ### Changed